requirements.txt
flax
optax
psgd_jax
einops
mosaicml-streaming
wandb
diffusers
jaxtyping
